+++ app/_components/BookViewModal/index.tsx
@@ -1,0 +1,32 @@
+import Typography from "@mui/material/Typography";
+import Box from "@mui/material/Box";
+import { Card } from "@/app/types";
+
+const style = {
+  position: "absolute" as "absolute",
+  top: "50%",
+  left: "50%",
+  transform: "translate(-50%, -50%)",
+  width: 400,
+  bgcolor: "background.paper",
+  border: "2px solid #000",
+  boxShadow: 24,
+  p: 4,
+};
+
+export const BookViewmodal = ({ card }: { card: Card }) => {
+  const { title, author, description } = card;
+  return (
+    <Box sx={style}>
+      <Typography id="modal-modal-title" variant="h6" component="h2">
+        {title}
+      </Typography>
+      <Typography id="modal-modal-description" sx={{ mt: 2 }}>
+        {author}
+      </Typography>
+      <Typography id="modal-modal-description" sx={{ mt: 2 }}>
+        {description}
+      </Typography>
+    </Box>
+  );
+};
